Switch visibility settings edit form and profile rendering to new system
This is how long a line can be if you were to type it in manually rather
This commit removes the old visibility settings fields (that used a form
alter) from the user settings page. This form is replaced by the profile
form rendered in the `visibility` form mode.

The profile was previously rendered by a view which caused field level
access checks to be ignored. To solve this the route is replaced with a
controller that simply renders the profile in the full view mode. This
ensures that field level access checks are correctly applied.

As part of the access check improvements the information stored about
fields in the field storage third party settings is changed slightly.
Rather than storing the field storage ID we now only store the field
name (for the corresponding visibility or value field). This makes
handling the value slightly easier and makes it applicable across both
FieldConfig and FieldStorage entities. *We don’t currently support
storing visibility settings on a different entity-type/-bundle than the
corresponding value field.*

The information storage change allows to make the FieldManager class
slightly more readable and provides clearer `isX` functions. Finally we
also generalise some of the entity type and bundle checks which in
theory makes this usable on other entity types or profile bundles too
(although the required hooks for field config/storage synchronisation)
still need to be implemented.

While testing both the user visibility settings editing and the profile
field display access checks it was discovered that some changes were
needed there too. Specifically returning Neutral from a field_access
hook will make the access check default to allowed. So in our case we
want to expressly allow or forbid rather than return neutral if we know
something about the access at all.