Fix update hook order after rebase
Quite a few changes have happened to update hooks so we move them all to
the end of the file and give them a name for Open Social 12.

We move around some existing hooks because they were accidentally moved
or duplicated during the rebase. This should bring the overall diff for
this branch as close to `main` as possible.