Issue #3275943 by alex.ksis: Reset cache of the commented entity.
In case of using social_ajax_comments feature for the comments form, the node caches are generated incorrectly, with no data for some users. Although the caches of the commented entity are cleared in the \Drupal\comment\Entity\Comment::postSave(), it does not work. As a temporary solution, we can clear the caches in the AjaxCommentsController class, after a new comment is saved.